adding dedicated admin operator test
